removed old landing page files and renamed Landing2 folder to Landing
