+++ client/web/src/apps/about/Routes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Switch, Route, Redirect } from 'react-router-dom'
-import Landing from 'components/pages/Landing2'
+import Landing from 'components/pages/Landing'
 import Privacy from 'components/pages/Privacy'
 
 const Routes = () => {
