+++ client/web/src/components/pages/Landing/Section2.js
